+++ absl/flags/tests/flags_numeric_bounds_test.py
@@ -14,14 +14,10 @@
 
 """Tests for lower/upper bounds validators for numeric flags."""
 
-from __future__ import absolute_import
-from __future__ import division
-from __future__ import print_function
-
+from unittest import mock
 from absl import flags
 from absl.flags import _validators
 from absl.testing import absltest
-import mock
 
 
 class NumericFlagBoundsTest(absltest.TestCase):